# sdk-gen/openapi.json
{
  "components": {
    "schemas": {
...
      "AppSettingsRead": {
        "properties": {
          "evaluate_temperature": {
            "title": "Evaluate Temperature",
            "type": "number"
          },
          "fallback_chain": {
            "items": {
              "type": "string"
            },
            "title": "Fallback Chain",
            "type": "array"
          },
          "has_anthropic_key": {
            "title": "Has Anthropic Key",
            "type": "boolean"
          },
          "has_deepl_key": {
            "title": "Has Deepl Key",
            "type": "boolean"
          },
          "has_openai_key": {
            "title": "Has Openai Key",
            "type": "boolean"
          },
          "has_openrouter_key": {
            "title": "Has Openrouter Key",
            "type": "boolean"
          },
          "ollama_host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ollama Host"
          },
          "openrouter_model": {
            "title": "Openrouter Model",
            "type": "string"
          },
          "primary_provider": {
            "title": "Primary Provider",
            "type": "string"
          },
          "translate_temperature": {
            "title": "Translate Temperature",
            "type": "number"
          }
        },
        "required": [
          "has_anthropic_key",
          "has_openai_key",
          "has_openrouter_key",
          "has_deepl_key",
          "openrouter_model",
          "primary_provider",
          "fallback_chain",
          "translate_temperature",
          "evaluate_temperature",
          "ollama_host"
        ],
        "title": "AppSettingsRead",
        "type": "object"
      },
      "AppSettingsUpdate": {
        "additionalProperties": false,
        "description": "Partial-update body \u2014 every field is optional.\n\nNote the distinction between \"field omitted\" (no-op) and \"field set to\nempty string\" (clear the encrypted column). Use ``exclude_unset=True``\non the server to make that split, not ``exclude_none``.",
        "properties": {
          "anthropic_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anthropic Api Key"
          },
          "deepl_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deepl Api Key"
          },
          "evaluate_temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluate Temperature"
          },
          "fallback_chain": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Chain"
          },
          "ollama_host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ollama Host"
          },
          "openai_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openai Api Key"
          },
          "openrouter_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Api Key"
          },
          "openrouter_model": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openrouter Model"
          },
          "primary_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Provider"
          },
          "translate_temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translate Temperature"
          }
        },
        "title": "AppSettingsUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProviderTestRequest": {
        "description": "Body for POST /app-settings/test. ``api_key`` is optional \u2014 when omitted,\nthe server tests the stored (decrypted) key for that provider so an admin can\nverify an already-saved credential. ``ollama_host`` is only used for ollama.",
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "ollama_host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ollama Host"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider"
        ],
        "title": "ProviderTestRequest",
        "type": "object"
      },
      "ProviderTestResult": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ProviderTestResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "ClaritiTMS",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/app-settings": {
      "get": {
        "operationId": "get_app_settings_api_v1_app_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get App Settings",
        "tags": [
          "App Settings"
        ]
      },
      "patch": {
        "operationId": "update_app_settings_api_v1_app_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettingsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Update App Settings",
        "tags": [
          "App Settings"
        ]
      }
    },
    "/api/v1/app-settings/test": {
      "post": {
        "description": "Validate a provider credential with a cheap authenticated call.\n\nTests the supplied ``api_key`` if given, otherwise the stored (decrypted)\nkey for that provider \u2014 so an admin can verify a key they're about to enter\nOR one already saved. Always returns 200 with ``{ok, error}``; a failed\ntest is data, not an HTTP error.",
        "operationId": "test_provider_api_v1_app_settings_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderTestResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Test Provider",
        "tags": [
          "App Settings"
        ]
      }
    },
...
  }
}